feat(style): Finish initial responsive adaptation
